fix: isolate cleanup evidence storage
